ITS3: cleanup code
