+++ Detectors/Upgrades/ITS3/alignment/include/ITS3Align/AlignmentHierarchy.h
@@ -13,14 +13,11 @@
 #define O2_ITS3_ALIGNMENT_HIERARCHY_H
 
 #include <memory>
-#include <compare>
-#include <type_traits>
-#include <map>
 #include <utility>
 #include <vector>
 #include <ostream>
 #include <string>
-#include <format>
+#include <map>
 #include <algorithm>
 
 #include <Eigen/Dense>
@@ -28,22 +25,13 @@
 #include <TGeoMatrix.h>
 #include <TGeoPhysicalNode.h>
 
+#include "ITS3Align/AlignmentLabel.h"
+#include "ITS3Align/AlignmentDOF.h"
+
 namespace o2::its3::align
 {
 using Matrix36 = Eigen::Matrix<double, 3, 6>;
 using Matrix66 = Eigen::Matrix<double, 6, 6>;
-
-// indices for rigid body parameters in LOC frame
-enum RigidBodyDOF : uint8_t {
-  TX = 0,
-  TY,
-  TZ,
-  RX,
-  RY,
-  RZ,
-  NDOF,
-};
-static constexpr const char* RigidBodyDOFNames[RigidBodyDOF::NDOF] = {"TX", "TY", "TZ", "RX", "RY", "RZ"};
 
 // return the rigid body derivatives
 // trk has be at in the measurment frame
@@ -63,144 +51,6 @@ auto getRigidBodyDerivatives(const auto& trk)
   return der;
 }
 
-class DOFSet
-{
- public:
-  enum class Type : uint8_t { RigidBody,
-                              Legendre };
-  virtual ~DOFSet() = default;
-  virtual Type type() const = 0;
-  int nDOFs() const { return static_cast<int>(mFree.size()); }
-  virtual std::string dofName(int idx) const = 0;
-  bool isFree(int idx) const { return mFree[idx]; }
-  void setFree(int idx, bool f) { mFree[idx] = f; }
-  void setAllFree(bool f) { std::fill(mFree.begin(), mFree.end(), f); }
-  int nFreeDOFs() const
-  {
-    int n = 0;
-    for (bool f : mFree) {
-      n += f;
-    }
-    return n;
-  }
-
- protected:
-  DOFSet(int n) : mFree(n, true) {}
-  std::vector<bool> mFree;
-};
-
-class RigidBodyDOFSet final : public DOFSet
-{
- public:
-  static constexpr int NDOF = RigidBodyDOF::NDOF;
-  RigidBodyDOFSet() : DOFSet(NDOF) {}
-  // mask: bitmask of free DOFs (bit i = DOF i is free)
-  explicit RigidBodyDOFSet(uint8_t mask) : DOFSet(NDOF)
-  {
-    for (int i = 0; i < NDOF; ++i) {
-      mFree[i] = (mask >> i) & 1;
-    }
-  }
-  Type type() const override { return Type::RigidBody; }
-  std::string dofName(int idx) const override { return RigidBodyDOFNames[idx]; }
-  uint8_t mask() const
-  {
-    uint8_t m = 0;
-    for (int i = 0; i < NDOF; ++i) {
-      m |= (uint8_t(mFree[i]) << i);
-    }
-    return m;
-  }
-};
-
-class LegendreDOFSet final : public DOFSet
-{
- public:
-  explicit LegendreDOFSet(int order) : DOFSet((order + 1) * (order + 2) / 2), mOrder(order) {}
-  Type type() const override { return Type::Legendre; }
-  int order() const { return mOrder; }
-  std::string dofName(int idx) const override
-  {
-    int i = 0;
-    while ((i + 1) * (i + 2) / 2 <= idx) {
-      ++i;
-    }
-    int j = idx - (i * (i + 1) / 2);
-    return std::format("L({},{})", i, j);
-  }
-
- private:
-  int mOrder;
-};
-
-class GlobalLabel
-{
-  // Millepede label is any positive integer [1....)
-  // Layout: DOF(5) | CALIB(1) | ID(22) | SENS(1) | DET(2) = 31 usable bits (MSB reserved, GBL uses signed int)
- public:
-  using T = uint32_t;
-  static constexpr int DOF_BITS = 5;   // bits 0-4
-  static constexpr int CALIB_BITS = 1; // bit 5: 0 = rigid body, 1 = calibration
-  static constexpr int ID_BITS = 22;   // bits 6-27
-  static constexpr int SENS_BITS = 1;  // bit 28
-  static constexpr int TOTAL_BITS = sizeof(T) * 8;
-  static constexpr int DET_BITS = TOTAL_BITS - (DOF_BITS + CALIB_BITS + ID_BITS + SENS_BITS) - 1; // one less bit since GBL uses int!
-  static constexpr T bitMask(int b) noexcept
-  {
-    return (T(1) << b) - T(1);
-  }
-  static constexpr int DOF_SHIFT = 0;
-  static constexpr T DOF_MAX = (T(1) << DOF_BITS) - T(1);
-  static constexpr T DOF_MASK = DOF_MAX << DOF_SHIFT;
-  static constexpr int CALIB_SHIFT = DOF_BITS;
-  static constexpr T CALIB_MAX = (T(1) << CALIB_BITS) - T(1);
-  static constexpr T CALIB_MASK = CALIB_MAX << CALIB_SHIFT;
-  static constexpr int ID_SHIFT = DOF_BITS + CALIB_BITS;
-  static constexpr T ID_MAX = (T(1) << ID_BITS) - T(1);
-  static constexpr T ID_MASK = ID_MAX << ID_SHIFT;
-  static constexpr int SENS_SHIFT = DOF_BITS + CALIB_BITS + ID_BITS;
-  static constexpr T SENS_MAX = (T(1) << SENS_BITS) - T(1);
-  static constexpr T SENS_MASK = SENS_MAX << SENS_SHIFT;
-  static constexpr int DET_SHIFT = DOF_BITS + CALIB_BITS + ID_BITS + SENS_BITS;
-  static constexpr T DET_MAX = (T(1) << DET_BITS) - T(1);
-  static constexpr T DET_MASK = DET_MAX << DET_SHIFT;
-
-  GlobalLabel(T det, T id, bool sens, bool calib = false)
-    : mID((((id + 1) & ID_MAX) << ID_SHIFT) |
-          ((det & DET_MAX) << DET_SHIFT) |
-          ((T(sens) & SENS_MAX) << SENS_SHIFT) |
-          ((T(calib) & CALIB_MAX) << CALIB_SHIFT))
-  {
-  }
-
-  /// produce the raw Millepede label for a given DOF index (rigid body: calib=0 in label)
-  constexpr T raw(T dof) const noexcept { return (mID & ~DOF_MASK) | ((dof & DOF_MAX) << DOF_SHIFT); }
-  constexpr int rawGBL(T dof) const noexcept { return static_cast<int>(raw(dof)); }
-
-  /// return a copy of this label with the CALIB bit set (for calibration DOFs on same volume)
-  GlobalLabel asCalib() const noexcept
-  {
-    GlobalLabel c{*this};
-    c.mID |= (T(1) << CALIB_SHIFT);
-    return c;
-  }
-
-  constexpr T id() const noexcept { return ((mID >> ID_SHIFT) & ID_MAX) - 1; }
-  constexpr T det() const noexcept { return (mID & DET_MASK) >> DET_SHIFT; }
-  constexpr bool sens() const noexcept { return (mID & SENS_MASK) >> SENS_SHIFT; }
-  constexpr bool calib() const noexcept { return (mID & CALIB_MASK) >> CALIB_SHIFT; }
-
-  std::string asString() const
-  {
-    return std::format("Det:{} Id:{} Sens:{} Calib:{}", det(), id(), sens(), calib());
-  }
-
-  constexpr auto operator<=>(const GlobalLabel&) const noexcept = default;
-
- private:
-  T mID{0};
-};
-
 class HierarchyConstraint
 {
  public:
@@ -219,8 +69,6 @@ class HierarchyConstraint
   std::vector<uint32_t> mLabels; // parameter labels
   std::vector<double> mCoeff;    // their coefficients
 };
-
-// --- AlignableVolume ---
 
 class AlignableVolume
 {
